pyproject.toml
[project]
name = "rag-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.152.0",
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
